crud reserva

+++ locadora/src/main/java/br/com/locadora/vmoura/dominio/repositorio/ClienteRepositorio.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import br.com.locadora.vmoura.dominio.entidade.Cliente;
 
-public interface ClienteRepositorio extends JpaRepository<Cliente, Long> {
+public interface ClienteRepositorio extends JpaRepository<Cliente, Integer> {
 
 }
